// frontend/lib/types.ts
export interface MonthlyCashFlowPoint {
  month: number;
  year: number;
  gross_rent: number;
  platform_fee: number;
  cleaning_fee: number;
  agency_fee: number;
  imu: number;
  tari: number;
  condominio: number;
  insurance: number;
  maintenance: number;
  utilities: number;
  rental_tax: number;
  mortgage_payment: number;
  mortgage_interest: number;
  mortgage_principal: number;
  net_cash_flow: number;
  cumulative_cash_flow: number;
  mortgage_balance: number;
  property_value: number;
}

export interface AnnualSummary {
  year: number;
  gross_rent: number;
  total_opex: number;
  rental_tax: number;
  mortgage_payment: number;
  mortgage_interest: number;
  mortgage_principal: number;
  net_cash_flow: number;
  cumulative_cash_flow: number;
  mortgage_balance: number;
  property_value: number;
  equity: number;
}

export interface PurchaseCostBreakdown {
  down_payment: number;
  registration_tax: number;
  vat: number;
  notary: number;
  agency: number;
  renovation: number;
  furnishing: number;
  total_initial_cash: number;
  loan_amount: number;
}

export interface AnalysisResult {
  summary: {
    initial_cash_required: number;
    loan_amount: number;
    monthly_mortgage_payment: number;
    break_even_month: number | null;
    year_1_net_cash_flow: number;
    year_5_net_cash_flow: number;
    total_roi_pct: number;
    cash_on_cash_return_pct: number;
    net_yield_pct: number;
    purchase_costs: PurchaseCostBreakdown;
  };
  monthly_series: MonthlyCashFlowPoint[];
  annual_series: AnnualSummary[];
}